+++ frontend/src/app/components/overview/cloud-summary/cloud-summary.less
@@ -1,0 +1,3 @@
+cloud-summary {
+    .column();
+}
