+++ frontend/src/app/components/overview/cluster-summary/cluster-summary.less
@@ -1,0 +1,3 @@
+cluster-summary {
+    display: block;
+}
