+++ frontend/src/app/components/overview/nodes-summary/nodes-summary.less
@@ -1,0 +1,7 @@
+nodes-summary {
+    .column();
+
+    .nodes-chart {
+        margin: auto;
+    }
+}
