+++ frontend/src/app/components/overview/overview-panel/overview-panel.less
@@ -16,10 +16,6 @@
         background-size: cover;
     }
 
-    .nodes-chart {
-        margin: auto;
-    }
-
     .overview-btn {
         font-size: @font-size4;
         width :100%;
@@ -28,13 +24,5 @@
 
     hr {
         border-width: 2px;
-    }
-
-    .usage-history-filter .toggle-btn {
-        min-width: 90px;
-    }
-
-    .prop-value {
-        text-align: right;
     }
 }
