+++ frontend/src/app/components/overview/storage-usage-summary/storage-usage-summary.less
@@ -1,3 +1,7 @@
 storage-usage-summary {
     .column();
+
+    .duration-chooser .toggle-btn {
+        min-width: 90px;
+    }
 }
